LPS-124846 Rename package com.liferay.fragment.renderer.collection.filter to com.liferay.fragment.renderer.collection.asset.categories.filter
