+++ modules/apps/fragment/fragment-renderer-collection-asset-categories-filter-impl/src/main/java/com/liferay/fragment/renderer/collection/asset/categories/filter/internal/constants/CollectionFilterFragmentRendererWebKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.fragment.renderer.collection.filter.internal.constants;
+package com.liferay.fragment.renderer.collection.asset.categories.filter.internal.constants;
 
 /**
  * @author Rubén Pulido
