+++ modules/apps/fragment/fragment-renderer-collection-asset-categories-filter-impl/src/main/resources/META-INF/resources/init.jsp
@@ -19,7 +19,7 @@
 <%@ taglib uri="http://liferay.com/tld/clay" prefix="clay" %><%@
 taglib uri="http://liferay.com/tld/react" prefix="react" %>
 
-<%@ page import="com.liferay.fragment.renderer.collection.filter.internal.display.context.CollectionFilterFragmentRendererDisplayContext" %><%@
+<%@ page import="com.liferay.fragment.renderer.collection.asset.categories.filter.internal.display.context.CollectionFilterFragmentRendererDisplayContext" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %>
 
 <%
