+++ modules/apps/fragment/fragment-renderer-collection-asset-categories-filter-impl/src/main/resources/com/liferay/fragment/renderer/collection/asset/categories/filter/internal/dependencies/configuration.json
